+++ ibllib/io/extractors/training_wheel.py
@@ -243,17 +243,21 @@ def extract_wheel_moves(re_ts, re_pos, display=False):
     """
     if len(re_ts.shape) == 1:
         assert re_ts.size == re_pos.size, 'wheel data dimension mismatch'
-        assert np.all(
-            np.diff(re_ts) > 0), 'wheel timestamps not monotonically increasing'
+        assert np.all(np.diff(re_ts) > 0), 'wheel timestamps not strictly increasing'
     else:
         _logger.debug('2D wheel timestamps')
+        if len(re_pos.shape) > 1:  # Ensure 1D array of positions
+            re_pos = re_pos.flatten()
+        # Linearly interpolate the times
+        x = np.arange(re_pos.size)
+        re_ts = np.interp(x, re_ts[:, 0], re_ts[:, 1])
 
     # Check the values and units of wheel position
     res = np.array([wh.ENC_RES, wh.ENC_RES / 2, wh.ENC_RES / 4])
     # min change in rad and cm for each decoding type
     # [rad_X4, rad_X2, rad_X1, cm_X4, cm_X2, cm_X1]
     min_change = np.concatenate([2 * np.pi / res, wh.WHEEL_DIAMETER * np.pi / res])
-    pos_diff = np.abs(np.ediff1d(re_pos)).min()
+    pos_diff = np.median(np.abs(np.ediff1d(re_pos)))
 
     # find min change closest to min pos_diff
     idx = np.argmin(np.abs(min_change - pos_diff))
@@ -323,6 +327,7 @@ def extract_first_movement_times(wheel_moves, trials, min_qt=None):
     flinch = abs(wheel_moves['peakAmplitude']) < THRESH
     all_move_onsets = wheel_moves['intervals'][:, 0]
     # Iterate over trials, extracting onsets approx. within closed-loop period
+    cwarn = 0
     for i, (t1, t2) in enumerate(zip(trials['goCue_times'] - min_qt,
                                      trials['feedback_times'])):
         if ~np.isnan(t2 - t1):  # If both timestamps defined
@@ -334,7 +339,9 @@ def extract_first_movement_times(wheel_moves, trials, min_qt=None):
                     first_move_onsets[i] = all_move_onsets[ids[i]]  # Save first large onset
                     is_final_movement[i] = ids[i] == trial_onset_ids[-1]  # Final move of trial
         else:  # Log missing timestamps
-            _logger.warning('no reliable times for trial id %i', i + 1)
+            cwarn += 1
+    if cwarn:
+        _logger.warning(f'no reliable goCue/Feedback times (both needed) for {cwarn} trials')
 
     return first_move_onsets, is_final_movement, ids[ids != -1]