+++ ibllib/qc/__init__.py
@@ -1,4 +1,0 @@
-from ibllib.qc.bpodqc_extractors import BpodQCExtractor
-from ibllib.qc.bpodqc_metrics import BpodQC
-from ibllib.qc.oneqc_metrics import ONEQC
-from ibllib.qc.extended_qc import ExtendedQC