+++ release_notes.md
@@ -1,4 +1,24 @@
-### Release Notes 1.5.0dev - (25/05/2020 - (EDIT THIS HEADER BEFORE RELEASE)
+### Release Notes 1.5.5
+-   ONE offline mode and cache dataset table to speed up reloading of large datasets (Olivier)
+-   ALF io naming conventions on loading objects (Miles)
+-   KS2 Matlab ephys pipeline tasks (Olivier)
+-   Support for running QC on biased and training sessions (Nico)
+-   metrics_df and passed_df properties in BpodQC obj for qcplots (Nico)
+-   Added missing unittest to stim_move_before_goCue metric (Nico)
+
+### Release Notes 1.5.4 - 29/07/2020 hotfix
+-   ibllib.pipes.training_preprocessing.TrainingAudio
+
+### Release Notes 1.5.3 - 28/07/2020
+-   ibllib.pipes.training_preprocessing.TrainingAudio: returns files for registration and proper status. (Olivier)
+-   ibllib.atlas: compute nearest region from probe trajectory (Mayo)
+    
+### Release Notes 1.5.2 - 23/07/2020
+-   Local server jobs:
+    -   fix wheel moves size mismatch extractor error
+    -   only look for raw_session.flag for ephys extraction to avoid race conditions
+
+### Release Notes 1.5.1 - 25/05/2020
 -  Ephys extraction:
     -   spike amplitudes in Volts
     -   added waveforms samples dataset to use Phy from Flatiron datasets